flaskblog/routes.py
from flask import render_template, url_for, flash , redirect, request, abort
from flaskblog import app,db, bcrypt, mail
from flaskblog.forms import (RegistrationForm, LoginForm,
                            UpdateAccountForm, PostForm, RequestResetForm, ResetPasswordForm)
from flaskblog.models import User,Post
from flask_login import login_user, current_user, logout_user, login_required
import secrets
from PIL import Image
import os
from flask_mail import Message






























