requirements.txt
asgiref==3.4.1
bcrypt==4.0.1
blinker==1.5
click==8.0.4
clipboard==0.0.4
colorama==0.4.5
cycler==0.11.0
dataclasses==0.8
Django==3.2.16
dnspython==2.2.1
email-validator==1.3.0
Flask==2.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.3.post0
gunicorn==20.1.0
idna==3.4
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.4
numpy==1.19.5
pandas==1.1.5
Pillow==8.4.0
psycopg2==2.9.5
pyparsing==3.0.9
pyperclip==1.8.2
python-dateutil==2.8.2
pytz==2022.4
scipy==1.5.4
seaborn==0.11.2
six==1.16.0
SQLAlchemy==1.4.42
sqlparse==0.4.3
typing_extensions==4.1.1
Werkzeug==2.0.3
WTForms==3.0.0
zipp==3.6.0
